realcalls: Allow to specify variadic arg functions
This is useful for wrapping functions such as fcntl, where the third
argument is optional and the type of it also dependant on the second
argument.

Unfortunately, just adding ", ..." to DLSYM_FUN doesn't work, because
"..." is already used in a different way for C++ template arguments, so
we can't directly use it to pass C-style variadic function arguments.

After thinking through a whole set of workarounds, the most painless
solution however came from @Stanley00 (thanks a bunch!) when asking in
the ##c++-general IRC channel on Freenode.

So the way this is done now is actually quite simple and instead of
providing the return type and arguments seperately in DlsymFun's
template arguments, we now pass the full type of the function we want to
wrap.

Signed-off-by: aszlig <aszlig@nix.build>