Add logging to various places around the code base
So far we only had FATAL and TRACK messages, but with just these
messages we're still a bit in the dark if we want to know what's going
on.

While the logging messages I just added still are not exhaustive enough,
especially for the DEBUG loglevel, we now at least have way more insight
into what's going on and I'm already using this to debug a few problems
I've seen in the wild.

Signed-off-by: aszlig <aszlig@nix.build>