Replace libsystemd usage with own implementation
The use of sd_listen_fds (or sd_listen_fds_with_names, which essentially
wraps it) has a few issues with the way we want to use it.

First of all, it checks whether the PID matches with the one in
LISTEN_PID and returns with failure if it doesn't.

Second, and probably the most annoying thing, is that it sets the
FD_CLOEXEC flag on all the file descriptors. This has the implication
that once a program calls execve(), the file descriptors are no longer
available.

If an application natively support systemd socket activation this mostly
isn't a problem, because usually those file descriptors are gathered in
one place. However, for ip2unix, we associate these file descriptors
based on a set of rules and we don't necessarily use all of these file
descriptors at once or even at all.

So for example if the program in question is wrapped by a shell script,
something as simple as close() could trigger gathering systemd file
descriptors, which then will cause the main program to fail because the
systemd socket fds are closed when the shell executes the main program.

Of course, the implementation right now still is quite dumb and it
should actually set the FD_CLOEXEC flag once the file descriptor is
actually in use and maybe even alter the LISTEN_FDS environment
variable, so that child processes won't try to re-use the same socket
file descriptor. I'm not sure however whether the latter is a good idea
and I need to investigate that.

Having our own implementation also means, that we now no longer need to
depend on systemd libraries and we only need the systemd-socket-activate
binary in order to run the tests associated with socket activation.

Dropping that dependency also has the advantage that it's more unlikely
that we get into a mutex deadlock if libsystemd uses one of our wrapped
C library calls.

We also no longer need to check whether sd_listen_fds_with_names is
supported, because that's already handled gracefully in our own
implementation.

Signed-off-by: aszlig <aszlig@nix.build>