socket: Improve handling of non-SOL_SOCKET options
So far when using setsockopt on levels other than SOL_SOCKET, we just
passed it to real::setsocket without further processing.

However, this is bad if we have an application that expects an IP socket
and subsequently fails if it was unable to send eg. an IPPROTO_TCP
option to a socket that has already been converted to AF_UNIX.

Now we only pass the setsockopt options directly to real::setsockopt
whenever we're certain that the socket option either needs to be cached
or the socket has already been converted to a Unix domain socket.

Signed-off-by: aszlig <aszlig@nix.build>