logging: Use syslog format if stderr is a socket
When running ip2unix as a systemd service, the logs printed by
journalctl look like this:

  ip2unix[16018]: ip2unix[16018] preload.cc:486:close DEBUG: ...

The process name with PID and the DEBUG label here are somewhat
redundant, because syslog already has levels for things like WARNING,
ERR, DEBUG, etc. and journalctl also highlights them accordingly.

So whenever stderr is a socket, we now just prefix the log message with
"<N>", where N is the syslog level number. If stderr is a socket and
doesn't understand this prefix, it will only produce uglier log messages
but it shouldn't[TM] break anything.

Originally I was searching for a way to detect whether we are running as
a systemd service and I found an answer on Stack Overflow[1] which I
however would consider a workaround and possibly unreliable because
we're not necessarily $MAINPID.

Instead, the idea about checking for socket came from @falconindy
(thanks!) over at the #systemd IRC channel on Freenode, which is a much
nicer solution than checking name of the parent process or even check
whether JOURNAL_STREAM is set (which would only work on systemd version
231 or later).

[1]: https://stackoverflow.com/a/39388659

Signed-off-by: aszlig <aszlig@nix.build>