Make __IP2UNIX_RULES serialisation more compact
This introduces a new serialiser, which should turn stack size limits
into a non-issue, because it's even more compact than the rule
specification format we have in -r.

The reason why this is a non-issue now is that *both* command line
arguments and environment variables contribute to the stack size. What
we're essentially doing is removing the arguments and *turn* it into an
environment variable that is more compact.

Our new serialiser is implemented using function overloading (with a
mixture of templates) and even though we pass pointer references to its
arguments it should be more stateless than if I'd use a struct or class.

While the serialiser itself (apart from maybe out of memory conditions)
can't fail, the deserialiser has plenty of occasions where a failure
would occur. So instead of using exceptions, the deserialiser basically
forms forms a chain of std::optional values, which either contain the
error or std::nullopt if it was successful.

Signed-off-by: aszlig <aszlig@nix.build>