preload: Return rule position in match_rule
This is in preparation to the upcoming systemd socket handling refactor
and right now the position isn't used at all.

However, we intend to use it for matching systemd sockets to rule
numbers and it also is useful for providing better log messages, as we
can include the rule number there.

Signed-off-by: aszlig <aszlig@nix.build>