systemd: Improve state handling for passed fds
So far, we re-used LISTEN_FDS and LISTEN_FDNAMES in order to assign the
systemd socket file descriptors to the rules. We did that using a queue
which holds all the file descriptors and another unordered_map that
holds just the named file descriptions.

However, this has a few problems:

  * In multi-process applications we might end up assigning the same
    file descriptors twice, if the process forks off and uses execve().
  * If an unassigned file descriptor overlaps with a mapped one, the
    mapped file descriptor could become unavailable or blocked once
    we're doing a mapped lookup.
  * Systemd file descriptors aren't properly closed after being in use
    and the program uses execve().

While the last item isn't fully addressed by this, the fix however is
rather easy to do and will be following soon.

The main thing we're addressing now are the details on how sockets get
associated to file descriptors passed by systemd. We now associate the
systemd file descriptors to the rules during initialisation, whereas
before we did it on every lookup.

Those differences are better explained with an example, so consider we
have the following environment variables:

  * LISTEN_FDS=5
  * LISTEN_FDNAMES=???:foo:bar:???:???

With the old and new method this buils down to the following:

  Without rule-aware processing   |       Rule-aware processing
 ---------------------------------|------------------------------------
 .-----.-----.-----.-----.-----.  |  .-----.-----.-----.  .-----.-----.
 | ??? | foo | bar | ??? | ??? |  |  | ??? | ??? | ??? |  | foo | bar |
 |  0  |  1  |  2  |  3  |  4  |  |  |  0  |  3  |  4  |  |  1  |  2  |
 `-----'-----'--.--'-----'-----'  |  `-----'-----'-----'  `-----'--.--'
                |                 |                                |
       acquire fdname "bar"       |                acquire fdname "bar"
                |                 |                                |
               \|/                |                               \|/
 .-----.-----.......-----.-----.  |  .-----.-----.-----.  .-----.......
 | ??? | foo |.....| ??? | ??? |  |  | ??? | ??? | ??? |  | foo |......
 |  0  |  1  |.....|  3  |  4  |  |  |  0  |  3  |  4  |  |  1  |......
 `-----'-----'.....`-----'-----'  |  `-----'-----'-----'  `-----'......
    |                             |     |
 acquire                          |  acquire
   \|/                            |    \|/
 .......-----.......-----.-----.  |  .......-----.-----.  .-----.......
 ......| foo |.....| ??? | ??? |  |  ......| ??? | ??? |  | foo |......
 ......|  1  |.....|  3  |  4  |  |  ......|  3  |  4  |  |  1  |......
 ......`-----'.....`-----'-----'  |  ......`-----'-----'  `-----'......
          |                       |           |
       acquire                    |        acquire
         \|/                      |          \|/
 ...................-----.-----.  |  .............-----.  .-----.......
 ..................| ??? | ??? |  |  ............| ??? |  | foo |......
 ..................|  3  |  4  |  |  ............|  4  |  |  1  |......
 ..................`-----'-----'  |  ............`-----'  `-----'......
          |                       |                          |
  acquire fdname "foo"            |                acquire fdname "foo"
          |                       |                         \|/
          `-----.                 |  .............-----.  .............
               \|/                |  ............| ??? |  .............
                                  |  ............|  4  |  .............
              ? ? ?               |  ............`-----'  .............
                                  |
             .|||||.              |               .|||||.
            ; x   x ;             |              ; O   O ;
            |   ^   |             |              |   ^   |
            `. .-. .'             |              `.`---'.'
              `---'               |                `---'

We now use our serialisation functions to serialise/deserialise to
"__IP2UNIX_SYSTEMD_FDS" and "__IP2UNIX_SYSTEMD_FDMAP" and change the
environment variables whenever we take one item off the map.

While at first glance this might look similar to LISTEN_FDS and
LISTEN_FDNAMES, it's values are different. LISTEN_FDNAMES only contains
file descriptor names, while __IP2UNIX_SYSTEMD_FDMAP is a map of rule
positions to file descriptors.

Same goes for __IP2UNIX_SYSTEMD_FDS, which differs in that it is a set
of file descriptors instead of just a single number representing the
available file descriptors. The reason is that we might have holes in
there (due to named file descriptors), so we can't use a single int (see
the diagram above, which illustrates this).

Signed-off-by: aszlig <aszlig@nix.build>