systemd: Set FD_CLOEXEC flag when removing fds
This is what sd_listen_fds() actually does in order to prevent sockets
leaking into subprocesses.

When switching to our own implementation we no longer set the flag and
we actually shouldn't do it when gathering the file descriptors because
at that point it's way too early and we'd lose the file descriptors even
on the tiniest shell wrapper.

Now, we're setting the flag as soon as we have removed the file
descriptor from the pool and associated it to a socket. So in the event
we replace the process the file descriptor gets properly closed now.

Signed-off-by: aszlig <aszlig@nix.build>