Add basic path globbing functionality
This adds a path globbing function similar to fnmatch(3) but a tad more
restrictive. For example we don't consider locales and we also treat
path names specially, similar to FNM_PATHNAME.

From fnmatch(3) on FNM_PATHNAME:

  If this flag is set, match a slash in string only with a slash in
  pattern and not by an asterisk (*) or a question mark (?)
  metacharacter, nor by a bracket expression ([]) containing a slash.

We also don't allow named character classes such as "[[:alnum:]]", which
are only really useful if we have to handle locales and we want to have
a simple and very explicit syntax.

The reason why I implemented this from scratch is because fnmatch(3)
doesn't support recursive globbing, which at the moment we do not
support either but we're going to in just a moment.

Before writing this from scratch I tried to find existing solutions but
most of them either didn't meet our requirements or they used a
recursive algorithm.

Another way to implement this would have been to transform the pattern
into a regular expression and use std::regex to do the matching. However
this would mean, that we need to escape the various special characters
which I found a tad too ugly, because we not only need to escape the
usual suspects (such as ".", "*", "+" and whatnot) but also make sure we
do not support named character classes.

We need this functionality for allowing to wrap existing Unix sockets,
eg. to do something like:

  ip2unix -r 'unix=/run/*.sock' ...

Of course, the 'unix' rule option name isn't final at this point, but
instead of matching an IP address or port we can then match existing
Unix sockets with this without the need to hardcode the full path.

Signed-off-by: aszlig <aszlig@nix.build>