preload: Unregister socket if no rule matches
We really don't want to track sockets where we know that no rule is
matched or that is explicitly ignored, because it could lead to errors
for sockets that would otherwise be fine.

For example if an application binds to a port that is ignored and later
tries to run accept() on it, the latter will fail because our socket
handler didn't register the socket as bound due to the fact that no rule
has matched.

I also needed to change our connector implementation for the added test,
because it now needs to pick an ephemeral port because a hardcoded port
might already be in use. This isn't the case for Nix builds as we're in
a private network namespace, but it might fail if you run the tests
without Nix.

Signed-off-by: aszlig <aszlig@nix.build>
(cherry picked from commit 4e71835feb03c0d2a0dddf23a732a9a1be62ab18)