serial: Add decoder and encoder for std::pair
We're going to use this very soon to serialise systemd fds, because
we'll need one additional information for each file descriptor, which is
whether the socket file descriptor points to an AF_INET/AF_INET6 socket
or not.

The serialisation is pretty much straigtforward and we use # and $ for
delimiting the pairs.

Signed-off-by: aszlig <aszlig@nix.build>