systemd: Support AF_INET/AF_INET6 socket units
So far whenever the systemd flag was used, we could only use the
ListenStream socket unit option with a path. If it would have been an IP
address, the address rewriting would work in the wrong way because
ip2unix was always assuming AF_UNIX sockets.

This is now no longer the case and ip2unix can now be used to turn any
application that doesn't support socket activation into one that does,
even though this isn't exactly 100% in the scope of the name "ip2unix".

The reason why I'm doing this anyway is because it greatly increases the
use cases for the application and even in my case I have a few services
which I so far used with Unix sockets only but as the infrastructure
gets bigger, IP support is getting inevitable and it's much nicer to
just change the ListenStream option and be done with it rather than
rewriting the whole service/socket file or even patching the application.

Signed-off-by: aszlig <aszlig@nix.build>