Rename socketPath rule file option to just "path"
The rule file options were defined earlier than the rule specification
options (via the -r command line flag) and when designing the rule
specification format, I tried to keep options as short as possible.

Unfortunately, this introduced inconsistencies between the naming of
options in both rule formats and the major offender is "socketPath" in
the rule file, where it is just "path" via the command line argument.

So when renaming, I didn't remove the old option but rather print a
deprecation warning instead to make sure we don't immediately break
things for users.

Signed-off-by: aszlig <aszlig@nix.build>