Rename socketActivation rule file option as well
In the previous commit, we renamed the socketPath option to reduce
differences between rule specifications via command line and options in
the rule file.

Another option is "socketActivation", which is just called "systemd" in
the CLI specification.

The name/semantics of this option however might still be subject to
change, depending on how we are going to rename the fdName option,
because setting eg. fdName to "foo" is equivalent to "systemd=foo" in
the CLI specification.

One example where we'd need to touch the "systemd" option again is if
we'll end up with nested options like this:

  systemd:
    enable: true
    fdName: foo

Signed-off-by: aszlig <aszlig@nix.build>