Revert merge of PR #4 (rulefile unification)
This reverts commit 31fa1ab01d575bff5571ded05e90773868974010, reversing
changes made to 54d96bba2ac9bd5a22e881409fca83e115acecdc.

The changes here were NOT intended to be merged into master, because
they're for the next minor version and there are still open questions as
detailed in the pull request.

Furthermore, I intended to merge this alongside the next release so the
README stays consistent with its implementation.