Fix segfault if sockaddr buffers are NULL
For example something like accept(fd, NULL, NULL) is perfectly valid and
it's actually the most obvious case when writing a very simple socket
client or server in C and you're not interested in the peer address.

It turns out that actually very few services in the wild are doing this
and they usually always pass the buffers for the peer address.

However, this is not the case for something like Boost.Asio and we will
segfault to death due to a null pointer dereference.

The reason why we didn't catch this earlier is that we're using Python
to test our implementation, which *always* passes those buffers to
accept (or accept4 to be more exact).

Now the problem is that it's not very easy to use accept(fd, NULL, NULL)
in Python, so I decided to add a small helper in C++ to our tests for
that. This messes up our test hierarchy a little bit, but we need to
clean up this mess someday anyway, so let's load another trash bag on
it... X-/

Nevertheless, the most important thing here is that we now also have a
test that covers this case, so embarrassing bugs like this one won't
happen again.

Signed-off-by: aszlig <aszlig@nix.build>
(cherry picked from commit ab25ef54a112a6ab3622d73ab8b0376a1b084bfa)