release.nix: Allow Hydra jobset override for badge
We have a new jobset called release-2.0.x, where the badge job will
fail, because it is trying to query the ip2unix/master jobset on Hydra.

Of course in order to make several badges for the releases, we need to
be able to override things via Hydra to avoid the need to backport
badge-specific code into older releases.

This means, that not only the purgeUrl is now residing in jobset
configuration, but also the badgeTitle, because having two badges with a
"builds" title isn't going to be very helpful.

Signed-off-by: aszlig <aszlig@nix.build>