Add GitHub Actions workflow
We already use Hydra to run our tests with Nix, but so far we did not
run any tests on other distributions or operating systems.

I already know that in principle we *should* support Mac OS X as well,
but I haven't really tested it. So adding a workflow for "macos" should
give us some feedback about whether it works.

The workflow file is completely off the reference documentation and I
haven't tested it locally since I don't run Docker (nor do I want to).

So if things go south, we probably need to fix things up in future
commits.

Signed-off-by: aszlig <aszlig@nix.build>