default.nix: Improve filter for src attribute
Since all the Nix expressions are only for building and testing, we
don't need them to be part of the source store path.

So what we do now is explicitly whitelisting all we need to build the
source and run the tests and blacklisting tests/vm, so that whenever we
change any of the Nix expressions that are not directly tied to the
source build, we shouldn't get a rebuild.

Signed-off-by: aszlig <aszlig@nix.build>