Add support for abstract sockets
Adding abstract namespace sockets has been requested a long time ago[1]
and while I also had a partial implementation laying around, it was also
so ugly that I didn't want to merge it.

The reason why it was ugly was that first of all, I added another
"abstract" field to struct Rule and second, I used a boolean is_abstract
all over the place to distinguish between socket_path and abstract.

While in the end this resulted in a functional implementation, it also
made the code messier and hard to extend.

In the years passing since the aforementioned issue, I had several
occasions where abstract sockets would have been useful, so today I
decided to rewrite the ugly implementation mentioned above.

The new implementation now simply abstracts away the socket_path field
into a SocketPath type, which then either is marked Type::ABSTRACT or
Type::FILESYSTEM.

This has the advantage that we only need to know the details about
SocketPath when we parse the rules and when we create the SockAddr,
everything in between is just changing types from std::string to
SocketPath.

[1]: https://github.com/nixcloud/ip2unix/issues/8

Signed-off-by: aszlig <aszlig@nix.build>
Closes: https://github.com/nixcloud/ip2unix/issues/8