Add a test specifically for rsession
When introducing replay of epoll_ctl calls, I only implemented a very
minimal test case. However since the problem seemed to persist according
to the bug reporter, I decided to implement a small test case that is
closer to the reporter's actual use case.

Unfortunately (or fortunately?), the test case succeeds on our part, so
we need to debug a bit more and/or improve the test case on our side.

Signed-off-by: aszlig <aszlig@nix.build>