globpath: Fix handling slashes in escaped cclass
We only allow to match slashes outside of character classes, because we
want to adhere to POSIX glob(7) as much as possible.

From the glob(7) manpage:

  Globbing is applied on each of the components of a pathname
  separately. A '/' in a pathname cannot be matched by a '?' or '*'
  wildcard, or by a range like "[.-0]". A range containing an explicit
  '/' character is syntactically incorrect. (POSIX requires that
  syntactically incorrect patterns are left unchanged.)

So essentially when we're scanning the pattern for slashes we no longer
need to take special care about character classes, since they're matched
verbatim.

Additionally, whenever we hit a slash in a character class, we now
immediately mark it as invalid and use literal matching.

Signed-off-by: aszlig <aszlig@nix.build>