release: Add -fsanitize tests
While not all of these tests are run with full integration tests, we at
least run as much as possible to catch errors such as the one we have
fixed earlier in globpath.

Unfortunately, -fsanitize=address and -fsanitize=thread need to have a
shared library loaded first, but since the meat of our project is a
library to be put in LD_PRELOAD, we're competing with lib[at]san.

Ideally, we'd have some code in our main executable which is able to
detect lib[at]san and then makes sure that the runtime loading behaviour
of the dynamic linker is done properly.

For now however, this is better than having nothing at all.

Signed-off-by: aszlig <aszlig@nix.build>