Fix code causing Clang warnings
For some reason, we had one occasion, where GCC (even thouh we use the
-Wzero-as-null-pointer-constant flag) didn't warn about the use of NULL
instead of nullptr, while Clang did emit a warning.

All of the other changes are to prevent -Wparentheses warnings and this
is more like a safety/convention warning rather than a real issue:

  warning: using the result of an assignment as a condition without parentheses
  note: place parentheses around the assignment to silence this warning

So let's say you have a condition like this:

  if (foo = somefun()) {
    ...
  }

If one really wanted to mean "foo == somefun()" instead but forgot to
add another equals sign, this could result in the program not behaving
as intended.

What -Wparentheses does is that it forces one to add extra parentheses
around the condition to explicitly state that it's an assignment, like
this:

  if ((foo = somefun())) {
    ...
  }

Despite fixing the Clang errors, ip2unix currently doesn't fully support
building with Clang yet, but that's something landing a future change
near you. :-)

Signed-off-by: aszlig <aszlig@nix.build>