Bump version to 2.1.3
While most of the changes are fixes, there is one controversial change
that might possibly warrant to move to a version 2.2, which is that we
now also wrap epoll_ctl.

However, since semver states that the minor version should be
incremented whenever there is a change in the _public_ API, I still
think that this belongs to a bugfix version, because the user-facing
part is still 100% the same as 2.1.2 and handling the epoll_ctl syscall
internally is something I'd consider a bugfix, since it unbreaks using
ip2unix for software using epoll.

Additionally in comparison to the last version bump in 678c01cfea1bfac5,
we're now *actually* incrementing the version, which I forgot and I just
updated the changelog.

Signed-off-by: aszlig <aszlig@nix.build>