Deprecate -F, --rules-data and --rules-file args
Since we already have a pretty concise rule specification format, there
really is no need to specify YAML or JSON *inline* via the
-F/--rules-data argument, since the -r argument is superior in every
way, given that YAML is whitespace/newline-sensitive.

Another point is that the --rules-file argument always sounded a bit
weird (is it "rules file" or "rule file"?) and just calling it --file is
much less awkward and also unambiguous since all that can be passed via
file are rules, similar to patterns for grep and its -f/--file option.

Right now, the -f/--file option still reads YAML and doesn't yet
recognise a newline separated list of rules. The implementation will be
a little bit trickier, since we want to keep the same command line
argument but allow for both formats to coexist and simply display a
deprecation warning for YAML.

The reason why I decided to remove YAML entirely is because when using
ip2unix in practice, I did not have even *one* occasion where such a
rule file would be helpful and instead just used the -r arguments.

Since the rule file was implemented first, the option names are still
unchanged to date, but when introducing the -r command line option, I
decided to go for shorter names instead.

This however is a bit confusing if you're used to -r and suddenly have
to use entirely different option names for the rule files, so it's just
one additional reason to remove it entirely.

Signed-off-by: aszlig <aszlig@nix.build>