Implement simplified rule file parser
When the file given via -f is not a YAML file, we now parse the file as
a newline-separated list of rule (-r) arguments with only a few extras
such as stripping whitespace to the left, ignoring comments (#) and
empty lines.

The reason why I didn't strip whitespace to the right of each line is
that some rule options allow spaces, for the example the "path" option
might contain a file name which ends with a space character.

We also perform a minimal check whether a given rule file is valid YAML
and then emit a deprecation warning. Since YAML is a format that allows
a wide variety of ways to specify certain data types, the check also
only treats the file argument as YAML if the root data type is a
sequence.

Originally I wanted to unify the rule file options with the CLI
arguments (#5), but some rule argument options are very difficult to
express in the YAML file.

One example would be the "port" option, which in the rule file has two
options - "port" and "portEnd" - while the rule arguments use a single
option with a range "port=start-end" (eg. port=1000-2000) instead.

This essentially means that we either *always* have some disparity
between rule args and rule files *or* we lose some conciseness for the
rule arguments.

So using the same format just makes sense and also will free us from the
yamlcpp dependency starting with ip2unix version 3.0.

Signed-off-by: aszlig <aszlig@nix.build>
Closes: https://github.com/nixcloud/ip2unix/pull/5