README: Remove references to YAML rule file format
We no longer support YAML and since it's scheduled for removal in
version 3.0, we should no longer provide documentation for it.

When adding deprecation warnings in 49b889b03eb5e39f4fc348966a989567d58,
I already removed a good chunk of the related documentation, but there
were still a few references left, which are hereby removed.

One reference that I still did not remove is the mention of the
IP2UNIX_RULE_FILE environment variable, which still doesn't throw a
deprecation warning yet. This will be fixed very soon though.

Signed-off-by: aszlig <aszlig@nix.build>