Deprecate IP2UNIX_RULE_FILE environment variable
This enviroment variable was used for debugging the initial prototype of
ip2unix, where we didn't have proper serialisation and also didn't have
a more concise rule specification format.

We also no longer support using YAML files for specifying the rules and
also no longer consider the environment variables that are passed to the
library to be part of its public interface.

Having another "configuration file entry point" to take care of and
write tests for is something we also want to avoid, so instead of
recognising newline-separated rules for IP2UNIX_RULE_FILE I'm hereby
removing it for good.

Of course, one advantage of using an environment variable for passing
the rules is that we don't hit limits imposed by the operating system
with a very large amount of rules.

If we want to avoid such limits we can still add a file reference to our
serialisation format instead.

Signed-off-by: aszlig <aszlig@nix.build>