README: Fix literal formatting for some paragraphs
The paragraph after the + in the systemd option shouldn't not be
indented, because AsciiDoc then treats it as a literal paragraph, which
is not what we want here.

So to make sure we don't accidentally run into this issue in the future,
I removed all the indentation of the descriptions of the flags/options.

Signed-off-by: aszlig <aszlig@nix.build>