README: Improve description for systemd use case
Using socket units for parallel startup of services might be one reason
to use them but it is by far the least compelling reason.

The most compelling reason (at least for me) is again: Better privilege
separation and lowering of the attack surface.

Another thing I introduced in version 2.1.0 is support for AF_INET and
AF_INET6 sockets for systemd, so adding this to the problem statement is
also a good idea since while it doesn't specifically fit the purpose of
ip2unix, it still could be very useful to a lot of people out there.

Signed-off-by: aszlig <aszlig@nix.build>