Revert "github: Remove workflow for Mac OS X"
This reverts commit c945f586c390bf591bc30cc67e67b3a63b6587db.

I initially removed the workflow because it was failing all the time and
always seeing commit messages with failed CI checks is not only pretty
unhelpful but we might miss other *relevant* failures because of that.

So this brings back the workflow for Darwin since we're now actually
working on at least trying to implement support for it.

The reason why I'm not adding this *after* adding support is that I
don't have a Mac OS X machine, so I have to purely relying on the GitHub
Actions workflow.

Signed-off-by: aszlig <aszlig@nix.build>